Airbnb London Analysis/requirements.txt
pandas==1.5.3
sqlalchemy==2.0.19
matplotlib==3.7.2
seaborn==0.12.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0